Rewrite for no invalidations
The goal here is to eliminate all invalidations without losing much
(hopefully any) functionality. Tests do NOT pass yet.
