--- src/tuples.jl
@@ -1,212 +1,0 @@
-
-"""
-    field_type(::Type{T}, f)
-
-Functionally equivalent to `fieldtype(T, f)` except `f` may be a static type.
-"""
-@inline field_type(::Type{T}, f::Union{Int,Symbol}) where {T} = fieldtype(T, f)
-@generated field_type(::Type{T}, ::StaticInt{N}) where {T,N} = fieldtype(T, N)
-@generated field_type(::Type{T}, ::StaticSymbol{S}) where {T,S} = fieldtype(T, S)
-
-@inline nstatic(::Val{N}) where {N} = ntuple(StaticInt, Val(N))
-
-@inline function invariant_permutation(@nospecialize(x::Tuple), @nospecialize(y::Tuple))
-    if y === x === nstatic(Val(nfields(x)))
-        return True()
-    else
-        return False()
-    end
-end
-
-permute(@nospecialize(x::Tuple), @nospecialize(perm::Val)) = permute(x, static(perm))
-@inline function permute(@nospecialize(x::Tuple), @nospecialize(perm::Tuple))
-    if invariant_permutation(nstatic(Val(nfields(x))), perm) === False()
-        return eachop(getindex, perm, x)
-    else
-        return x
-    end
-end
-
-"""
-    eachop(op, args...; iterator::Tuple{Vararg{StaticInt}}) -> Tuple
-
-Produces a tuple of `(op(args..., iterator[1]), op(args..., iterator[2]),...)`.
-"""
-@inline function eachop(op::F, itr::Tuple{T,Vararg{Any}}, args::Vararg{Any}) where {F,T}
-    (op(args..., first(itr)), eachop(op, Base.tail(itr), args...)...)
-end
-eachop(::F, ::Tuple{}, args::Vararg{Any}) where {F} = ()
-
-"""
-    eachop_tuple(op, arg, args...; iterator::Tuple{Vararg{StaticInt}}) -> Type{Tuple}
-
-Produces a tuple type of `Tuple{op(arg, args..., iterator[1]), op(arg, args..., iterator[2]),...}`.
-Note that if one of the arguments passed to `op` is a `Tuple` type then it should be the first argument
-instead of one of the trailing arguments, ensuring type inference of each element of the tuple.
-"""
-eachop_tuple(op, itr, arg, args...) = _eachop_tuple(op, itr, arg, args)
-@generated function _eachop_tuple(op, ::I, arg, args::A) where {A,I}
-    t = Expr(:curly, Tuple)
-    narg = length(A.parameters)
-    for p in I.parameters
-        call_expr = Expr(:call, :op, :arg)
-        if narg > 0
-            for i in 1:narg
-                push!(call_expr.args, :(getfield(args, $i)))
-            end
-        end
-        push!(call_expr.args, :(StaticInt{$(p.parameters[1])}()))
-        push!(t.args, call_expr)
-    end
-    Expr(:block, Expr(:meta, :inline), t)
-end
-
-#=
-    find_first_eq(x, collection::Tuple)
-
-Finds the position in the tuple `collection` that is exactly equal (i.e. `===`) to `x`.
-If `x` and `collection` are static (`is_static`) and `x` is in `collection` then the return
-value is a `StaticInt`.
-=#
-@generated function find_first_eq(x::X, itr::I) where {X,N,I<:Tuple{Vararg{Any,N}}}
-    # we avoid incidental code gen when evaluated a tuple of known values by iterating
-    #  through `I.parameters` instead of `known(I)`.
-    index = ifelse(known(X) === nothing, nothing, findfirst(==(X), I.parameters))
-    if index === nothing
-        :(Base.Cartesian.@nif $(N + 1) d->(x == getfield(itr, d)) d->(d) d->(nothing))
-    else
-        :($(static(index)))
-    end
-end
-
-"""
-  reduce_tup(f::F, inds::Tuple{Vararg{Any,N}}) where {F,N}
-
-An optimized `reduce` for tuples. `Base.reduce`'s `afoldl` will often not inline.
-Additionally, `reduce_tup` attempts to order the reduction in an optimal manner.
-
-```julia
-julia> using StaticArrays, ArrayInterface, BenchmarkTools
-
-julia> rsum(v::SVector) = ArrayInterface.reduce_tup(+, v.data)
-rsum (generic function with 2 methods)
-
-julia> for n ∈ 2:16
-           @show n
-           v = @SVector rand(n)
-           s1 = @btime  sum(\$(Ref(v))[])
-           s2 = @btime rsum(\$(Ref(v))[])
-       end
-n = 2
-  0.863 ns (0 allocations: 0 bytes)
-  0.863 ns (0 allocations: 0 bytes)
-n = 3
-  0.862 ns (0 allocations: 0 bytes)
-  0.863 ns (0 allocations: 0 bytes)
-n = 4
-  0.862 ns (0 allocations: 0 bytes)
-  0.862 ns (0 allocations: 0 bytes)
-n = 5
-  1.074 ns (0 allocations: 0 bytes)
-  0.864 ns (0 allocations: 0 bytes)
-n = 6
-  0.864 ns (0 allocations: 0 bytes)
-  0.862 ns (0 allocations: 0 bytes)
-n = 7
-  1.075 ns (0 allocations: 0 bytes)
-  0.864 ns (0 allocations: 0 bytes)
-n = 8
-  1.077 ns (0 allocations: 0 bytes)
-  0.865 ns (0 allocations: 0 bytes)
-n = 9
-  1.081 ns (0 allocations: 0 bytes)
-  0.865 ns (0 allocations: 0 bytes)
-n = 10
-  1.195 ns (0 allocations: 0 bytes)
-  0.867 ns (0 allocations: 0 bytes)
-n = 11
-  1.357 ns (0 allocations: 0 bytes)
-  1.400 ns (0 allocations: 0 bytes)
-n = 12
-  1.543 ns (0 allocations: 0 bytes)
-  1.074 ns (0 allocations: 0 bytes)
-n = 13
-  1.702 ns (0 allocations: 0 bytes)
-  1.077 ns (0 allocations: 0 bytes)
-n = 14
-  1.913 ns (0 allocations: 0 bytes)
-  0.867 ns (0 allocations: 0 bytes)
-n = 15
-  2.076 ns (0 allocations: 0 bytes)
-  1.077 ns (0 allocations: 0 bytes)
-n = 16
-  2.273 ns (0 allocations: 0 bytes)
-  1.078 ns (0 allocations: 0 bytes)
-```
-
-More importantly, `reduce_tup(_pick_range, inds)` often performs better than `reduce(_pick_range, inds)`.
-```julia
-julia> using ArrayInterface, BenchmarkTools, Static
-
-julia> inds = (Base.OneTo(100), 1:100, 1:static(100))
-(Base.OneTo(100), 1:100, 1:static(100))
-
-julia> @btime reduce(ArrayInterface._pick_range, \$(Ref(inds))[])
-  6.405 ns (0 allocations: 0 bytes)
-Base.Slice(static(1):static(100))
-
-julia> @btime ArrayInterface.reduce_tup(ArrayInterface._pick_range, \$(Ref(inds))[])
-  2.570 ns (0 allocations: 0 bytes)
-Base.Slice(static(1):static(100))
-
-julia> inds = (Base.OneTo(100), 1:100, 1:UInt(100))
-(Base.OneTo(100), 1:100, 0x0000000000000001:0x0000000000000064)
-
-julia> @btime reduce(ArrayInterface._pick_range, \$(Ref(inds))[])
-  6.411 ns (0 allocations: 0 bytes)
-Base.Slice(static(1):100)
-
-julia> @btime ArrayInterface.reduce_tup(ArrayInterface._pick_range, \$(Ref(inds))[])
-  2.592 ns (0 allocations: 0 bytes)
-Base.Slice(static(1):100)
-
-julia> inds = (Base.OneTo(100), 1:100, 1:UInt(100), Int32(1):Int32(100))
-(Base.OneTo(100), 1:100, 0x0000000000000001:0x0000000000000064, 1:100)
-
-julia> @btime reduce(ArrayInterface._pick_range, \$(Ref(inds))[])
-  9.048 ns (0 allocations: 0 bytes)
-Base.Slice(static(1):100)
-
-julia> @btime ArrayInterface.reduce_tup(ArrayInterface._pick_range, \$(Ref(inds))[])
-  2.569 ns (0 allocations: 0 bytes)
-Base.Slice(static(1):100)
-```
-"""
-@generated function reduce_tup(f::F, inds::Tuple{Vararg{Any,N}}) where {F,N}
-    q = Expr(:block, Expr(:meta, :inline, :propagate_inbounds))
-    if N == 1
-        push!(q.args, :(inds[1]))
-        return q
-    end
-    syms = Vector{Symbol}(undef, N)
-    i = 0
-    for n ∈ 1:N
-        syms[n] = iₙ = Symbol(:i_, (i += 1))
-        push!(q.args, Expr(:(=), iₙ, Expr(:ref, :inds, n)))
-    end
-    W =  1 << (8sizeof(N) - 2 - leading_zeros(N))
-    while W > 0
-        _N = length(syms)
-        for _ ∈ 2W:W:_N
-            for w ∈ 1:W
-                new_sym = Symbol(:i_, (i += 1))
-                push!(q.args, Expr(:(=), new_sym, Expr(:call, :f, syms[w], syms[w+W])))
-                syms[w] = new_sym
-            end
-            deleteat!(syms, 1+W:2W)
-        end
-        W >>>= 1
-    end
-    q
-end
-
